refactor: restructure pgq
