+++ pkg/worker/job.go
@@ -13,24 +13,42 @@ import (
 const ctxCancellationBackoff = 3 * time.Second
 
 const (
-	StatusDone    = "DONE"
-	StatusPanic   = "PANIC"
-	StatusFailed  = "FAILED"
+	// StatusDone indicates the Job is successfully finished.
+	StatusDone = "DONE"
+
+	// StatusPanic indicates there was a panic during job-execution.
+	// This is a terminal status and will NOT be retried.
+	StatusPanic = "PANIC"
+
+	// StatusFailed indicates job failed to succeed even after retries.
+	// This is a terminal status and will NOT be retried.
+	StatusFailed = "FAILED"
+
+	// StatusPending indicates at-least 1 attempt is still pending.
 	StatusPending = "PENDING"
+)
+
+var (
+	ErrInvalidJob  = errors.New("job is not valid")
+	ErrKindExists  = errors.New("handler for given kind exists")
+	ErrUnknownKind = errors.New("job kind is invalid")
 )
 
 // Job represents the specification for async processing and also maintains
 // the progress so far.
 type Job struct {
-	ID        string    `json:"id"`
-	Kind      string    `json:"kind"`
+	// Specification of the job.
+	ID      string    `json:"id"`
+	Kind    string    `json:"kind"`
+	RunAt   time.Time `json:"run_at"`
+	Payload []byte    `json:"payload"`
+
+	// Internal metadata.
 	Status    string    `json:"status"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
-	RunAt     time.Time `json:"run_at"`
-	Payload   []byte    `json:"payload"`
 
-	// Result generated by job execution.
+	// Execution information.
 	Result        []byte    `json:"result,omitempty"`
 	AttemptsDone  int64     `json:"attempts_done"`
 	LastAttemptAt time.Time `json:"last_attempt_at,omitempty"`
@@ -55,6 +73,13 @@ type JobQueue interface {
 // job should be retried or marked with StatusFailed accordingly.
 // Refer RetryableError for expected behaviour on error.
 type JobFn func(ctx context.Context, job Job) ([]byte, error)
+
+// RetryableError can be returned by a JobFn to instruct the worker to attempt
+// retry after time specified by the RetryAfter field.
+type RetryableError struct {
+	Cause      error
+	RetryAfter time.Duration
+}
 
 func (j *Job) Sanitise() error {
 	now := time.Now()
@@ -84,6 +109,8 @@ func (j *Job) Sanitise() error {
 	return nil
 }
 
+// Attempt attempts to safely invoke `fn` for this job. Handles success, failure
+// and panic scenarios and updates the job with result in-place.
 func (j *Job) Attempt(ctx context.Context, now time.Time, fn JobFn) {
 	defer func() {
 		if v := recover(); v != nil {
@@ -119,4 +146,8 @@ func (j *Job) Attempt(ctx context.Context, now time.Time, fn JobFn) {
 			j.Status = StatusDone
 		}
 	}
+}
+
+func (re *RetryableError) Error() string {
+	return fmt.Sprintf("retryable-error: %v", re.Cause)
 }
